+++ packages/roosterjs-content-model-plugins/lib/tableEdit/OnTableEditorCreatedCallback.ts
@@ -1,4 +1,4 @@
-import type { TableEditFeatureName } from './editors/features/TableEditFeature';
+import type { TableEditFeatureName } from './editors/features/TableEditFeatureName';
 
 /**
  * Optional callback when creating a TableEditPlugin, allows to customize the Selectors element as required.
