+++ packages/roosterjs-content-model-plugins/lib/tableEdit/editors/features/TableEditFeatureName.ts
@@ -1,0 +1,10 @@
+/**
+ * Names of table edit features
+ */
+export type TableEditFeatureName =
+    | 'HorizontalTableInserter'
+    | 'VerticalTableInserter'
+    | 'TableMover'
+    | 'TableResizer'
+    | 'TableSelector'
+    | 'CellResizer';
